tcp related

+++ src/Phwang/Protocols/FabricFrontEndProtocolClass.java
@@ -9,7 +9,7 @@
 package Phwang.Protocols;
 
 public class FabricFrontEndProtocolClass {
-    public static final int LINK_MGR_PROTOCOL_TRANSPORT_PORT_NUMBER = 8006;
+    public static final short LINK_MGR_PROTOCOL_TRANSPORT_PORT_NUMBER = 8006;
 
     public static final int AJAX_MAPING_ID_SIZE = 3;
     public static final int WEB_FABRIC_PROTOCOL_NAME_LIST_TAG_SIZE = 3;
